improve ui to factor repeats - show user when the next due date is - show user when bill was paid

+++ app/src/main/java/app/monkpad/billmanager/presentation/newbill/NewBillFragment.kt
@@ -23,7 +23,7 @@ import com.google.android.material.dialog.MaterialAlertDialogBuilder
 class NewBillFragment : Fragment() {
     private lateinit var binding: FragmentNewBillBinding
     private val args: NewBillFragmentArgs by navArgs()
-    private var repeat: Int? = null
+    private var repeat: Long? = null
 
     private val viewModel: NewBillViewModel by activityViewModels{
         BillManagerViewModelFactory
